Update dependencies needed to build updated juju2
Since the interface needed for github.com/juju/txn differs between juju1
and juju2, revert the juju2 changes that rely on it.

Doesn't build at this point - the juju1 export code is broken by the
description changes, and baseClientCommand is broken by changes to
modelcmd - but I wanted those changes to be in a separate
commit.