Update baseClientCommand to use new modelcmd features
These are needed to get the macaroons to open a connection to the target
controller from the remote command.

upgrade-agents now gets to where it used to - gets macaroon from the
target controller and sends it to the remote command, which can use it
to download the binaries and unpack them.