Implement upgrade-agents command
Pushes the agent binaries and the Python upgrade script out to the
machines. The script captures the current agent configs and symlinks for
rollback, and moves the new tools into place.

The script also has a rollback option that will be used by the
rollback-agents command.

It doesn't yet update the agent configs or do anything very sensible
with the output from the script.