Start LXD containers up after migration
After migrating LXC containers to LXD,
start the LXD containers up so that
other commands can SSH into them to
run the remaining parts of the upgrade,
such as upgrade-agents. After starting
the LXD containers, the Juju agents are
stopped.

The "serviceCall" and "serviceCommand"
functions are consolidated into a new
function, agentServiceCommand. This
function will return an error if any
of the executions exit with a non-zero
code. Failed executions will have their
output written to stderr, prefixed with
the machine ID.