Updates to lxc-to-lxd script
We update the script to ignore the following config:
 - lxc.tty
 - lxc.devttydir
 - lxc.pivotdir
 - lxc.hook.clone

The first two should be irrelevant for cloud-based
software, and aren't supported by LXD anyway.

The lxc.pivotdir config is ignored by LXC now, and
is not important.

The lxc.hook.clone config is used by the Ubuntu
Cloud image templates for LXC, and is irrelevant
because we don't clone LXD containers in Juju 2.x.