Handle containers as SSH targets
Proxy SSH connections to containers
through their hosts.

Refactor parallelCall, renaming it
to parallelExec, and dropping the
irrelevant details of DistResult.
It now takes a list of targets
(machine/host addresses), and a
command to execute on each. The
callers mix the results with
machine details as needed.

Also, Stop swallowing errors when
making "service" command calls. We
are still just logging the output of
failed "service" command executions,
but we no longer swallow the outright
SSH transport failures.