Add update-maas-agentname command
In Juju 1.25, we assigned a new UUID specifically
for the MAAS agent_name to associate nodes to
Juju environments. In Juju 2.x, we use the model
UUID for the agent_name.

We Introduce a new command that will update the
maas-agent-name config attribute in the Juju
1.25 environ config, instructing the user to
run a command on the MAAS region controller
to update the MAAS database. This is necessary
as there is no API to change the agent_name
for a node once it has been acquired.