Exclude invalid relation units from export
If there's a subordinate application related to multiple principal ones,
in Juju 1.25 there'll be relation scopes for units that shouldn't be
involved in the relation - eg:

ubuntu/0
  telegraf/0
ntpmaster/0
  telegraf/1

There'll be a relation scope for telegraf/1 for the ubuntu-telegraf
relation. Exclude this from the export, since in Juju 2.2.3 it wouldn't
be there, and things get a bit confused if it is.

Ported this change back from Juju 2.2.3.