Make the names match what AdoptResources would make them
For upgrades on Openstack the security group controller update that
AdoptResources won't work for up (with nova networking) - it will crash
because it considers all of the groups even if they're not juju ones.

Make the group names match what AdoptResources should do. In activate
we detect that specific error and allow it (but we still call
AdoptResources for the other changes it makes).