Get rid of erroneous juju/juju imports
These were causing a problem where the juju/utils package was included
twice - once directly, and once via the juju/juju vendor
directory (because juju/juju was checked out to a 2.5-based
branch). That caused a panic because juju/utils has an init() that
registers a transport for the file: protocol, and the init was running
twice.